// app/src/exercise/exercise.dto.ts
export class ExerciseDto {
    name?: string;
    description?: string;
    link?: string;
    pictureLink?: string;
    gifLink?: string;
    lang?: string;
    forGames?: string[];
    classifierId?: string;
    exerciseType?: string;
    commandRight?: string;
    commandLeft?: string;
    bodyPart?: string;
    displayForPatient?: boolean;
    displayForSpecialist?: boolean;
    translation?: string;
    specialist?: string;
    phonePlacement?: string;
}